# backend/server.js
// Production-ready SaaS server with MySQL and full CRUD operations
require('dotenv').config();
const express = require('express');
const cors = require('cors');
const jwt = require('jsonwebtoken');
const { v4: uuidv4 } = require('uuid');
const { initDatabase, tenantDb, userDb, healthCheck } = require('./database');
const { 
  SUBSCRIPTION_PLANS, 
  initSubscriptionTables, 
  createDefaultSubscriptions,
  subscriptionManager, 
  requireFeature, 
  checkUsageLimit 
} = require('./subscription');

const app = express();
const PORT = process.env.PORT || 3001;

// Middleware
app.use(cors());
app.use(express.json());

// Request logging middleware
app.use((req, res, next) => {
  console.log(`${new Date().toISOString()} - ${req.method} ${req.path}`);
  next();
});

// Enhanced JWT middleware with better error handling
const authenticateToken = async (req, res, next) => {
  const authHeader = req.headers['authorization'];
  const token = authHeader && authHeader.split(' ')[1];

  if (!token) {
    return res.status(401).json({ 
      error: { 
        code: 'NO_TOKEN', 
        message: 'Access token required' 
      } 
    });
  }

  try {
    let decoded;
    try {
      // Try standard JWT decode first
      decoded = jwt.decode(token);
    } catch (e) {
      // Standard JWT decode failed
    }
    
    if (!decoded) {
      // If that fails, try manual decode for our mock format
      const parts = token.split('.');
      if (parts.length === 3) {
        try {
          const payloadString = Buffer.from(parts[1], 'base64').toString();
          decoded = JSON.parse(payloadString);
        } catch (decodeError) {
          console.log('Manual decode failed:', decodeError.message);
        }
      }
    }
    
    if (!decoded || !decoded.tenantId) {
      return res.status(403).json({ 
        error: { 
          code: 'INVALID_TOKEN', 
          message: 'Invalid token or missing tenant' 
        } 
      });
    }

    // Verify tenant exists
    const tenant = await tenantDb.getById(decoded.tenantId);
    if (!tenant || tenant.status !== 'active') {
      return res.status(403).json({ 
        error: { 
          code: 'TENANT_INACTIVE', 
          message: 'Tenant not found or inactive' 
        } 
      });
    }

    req.user = {
      tenantId: decoded.tenantId,
      userId: decoded.sub,
      roles: decoded.roles || ['user'],
      email: decoded.email,
      name: decoded.name
    };

    req.tenant = tenant;
    next();
  } catch (error) {
    console.log('JWT decode error:', error.message);
    return res.status(403).json({ 
      error: { 
        code: 'TOKEN_ERROR', 
        message: 'Token verification failed' 
      } 
    });
  }
};

// Role-based authorization middleware
const requireRole = (roles) => {
  return (req, res, next) => {
    if (!req.user || !req.user.roles.some(role => roles.includes(role))) {
      return res.status(403).json({
        error: {
          code: 'INSUFFICIENT_PERMISSIONS',
          message: `Required roles: ${roles.join(', ')}`
        }
      });
    }
    next();
  };
};

// =============================================================================
// SUBSCRIPTION MANAGEMENT ENDPOINTS
// =============================================================================

// Get available subscription plans
app.get('/api/v1/plans', async (req, res) => {
  try {
    const plans = Object.values(SUBSCRIPTION_PLANS).map(plan => ({
      id: plan.id,
      name: plan.name,
      price: plan.price,
      currency: plan.currency,
      billing_cycle: plan.billing_cycle,
      features: plan.features
    }));
    
    res.json(plans);
  } catch (error) {
    console.error('Error fetching plans:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch subscription plans'
      }
    });
  }
});

// Get current subscription
app.get('/api/v1/subscription', authenticateToken, async (req, res) => {
  try {
    const subscription = await subscriptionManager.getSubscription(req.user.tenantId);
    if (!subscription) {
      return res.status(404).json({
        error: {
          code: 'SUBSCRIPTION_NOT_FOUND',
          message: 'No active subscription found'
        }
      });
    }
    
    res.json(subscription);
  } catch (error) {
    console.error('Error fetching subscription:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch subscription'
      }
    });
  }
});

// Get usage statistics
app.get('/api/v1/usage', authenticateToken, async (req, res) => {
  try {
    const usageStats = await subscriptionManager.getUsageStats(req.user.tenantId);
    const subscription = await subscriptionManager.getSubscription(req.user.tenantId);
    
    res.json({
      usage: usageStats,
      plan: subscription ? subscription.plan : null,
      limits: subscription ? subscription.plan.limits : {}
    });
  } catch (error) {
    console.error('Error fetching usage:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch usage statistics'
      }
    });
  }
});

// Change subscription plan
app.post('/api/v1/subscription/change', authenticateToken, requireRole(['admin']), async (req, res) => {
  try {
    const { plan_id } = req.body;
    
    if (!plan_id || !SUBSCRIPTION_PLANS[plan_id]) {
      return res.status(400).json({
        error: {
          code: 'INVALID_PLAN',
          message: 'Invalid subscription plan'
        }
      });
    }
    
    const newSubscription = await subscriptionManager.changeSubscription(
      req.user.tenantId, 
      plan_id, 
      req.user.userId
    );
    
    res.json({
      message: 'Subscription updated successfully',
      subscription: newSubscription
    });
  } catch (error) {
    console.error('Error changing subscription:', error);
    res.status(500).json({
      error: {
        code: 'UPDATE_ERROR',
        message: 'Failed to update subscription'
      }
    });
  }
});

// =============================================================================
// FEATURE-GATED ENDPOINTS (Examples)
// =============================================================================

// Advanced analytics (Premium+ feature)
app.get('/api/v1/analytics', authenticateToken, requireFeature('advanced_analytics'), async (req, res) => {
  try {
    // Track usage
    await subscriptionManager.trackUsage(req.user.tenantId, 'api_calls');
    
    // Mock analytics data
    const analytics = {
      users_growth: [
        { month: 'Jan', users: 10 },
        { month: 'Feb', users: 15 },
        { month: 'Mar', users: 22 }
      ],
      revenue: {
        current_month: 1250.00,
        last_month: 980.00,
        growth: 27.5
      },
      top_features: [
        { name: 'User Management', usage: 85 },
        { name: 'Analytics', usage: 62 },
        { name: 'Integrations', usage: 41 }
      ]
    };
    
    res.json(analytics);
  } catch (error) {
    console.error('Error fetching analytics:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch analytics'
      }
    });
  }
});

// Team collaboration (Basic+ feature)
app.get('/api/v1/teams', authenticateToken, requireFeature('team_collaboration'), async (req, res) => {
  try {
    await subscriptionManager.trackUsage(req.user.tenantId, 'api_calls');
    
    // Mock team data
    const teams = [
      { id: 'team_1', name: 'Development Team', members: 5 },
      { id: 'team_2', name: 'Marketing Team', members: 3 }
    ];
    
    res.json(teams);
  } catch (error) {
    console.error('Error fetching teams:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch teams'
      }
    });
  }
});

// File upload with usage limits
app.post('/api/v1/upload', authenticateToken, checkUsageLimit('file_uploads'), async (req, res) => {
  try {
    // Track the upload
    await subscriptionManager.trackUsage(req.user.tenantId, 'file_uploads');
    
    // Mock file upload
    const fileId = `file_${uuidv4()}`;
    
    res.json({
      message: 'File uploaded successfully',
      file_id: fileId,
      remaining_uploads: req.usageLimit.remaining - 1
    });
  } catch (error) {
    console.error('Error uploading file:', error);
    res.status(500).json({
      error: {
        code: 'UPLOAD_ERROR',
        message: 'Failed to upload file'
      }
    });
  }
});

// =============================================================================
// TENANT MANAGEMENT ENDPOINTS (Super Admin only for demo)
// =============================================================================

// Get all tenants (with pagination and search)
app.get('/api/v1/admin/tenants', async (req, res) => {
  try {
    const { page = 1, limit = 10, search = '' } = req.query;
    const result = await tenantDb.getAll(parseInt(page), parseInt(limit), search);
    res.json(result);
  } catch (error) {
    console.error('Error fetching tenants:', error);
    res.status(500).json({ 
      error: { 
        code: 'FETCH_ERROR', 
        message: 'Failed to fetch tenants' 
      } 
    });
  }
});

// Create new tenant
app.post('/api/v1/admin/tenants', async (req, res) => {
  try {
    const { name, domain, plan = 'free', settings = {} } = req.body;
    
    if (!name || !domain) {
      return res.status(400).json({
        error: {
          code: 'MISSING_FIELDS',
          message: 'Name and domain are required'
        }
      });
    }

    const tenantId = `tenant_${uuidv4().replace(/-/g, '').substring(0, 8)}`;
    
    const tenant = await tenantDb.create({
      id: tenantId,
      name,
      domain,
      plan,
      settings
    }, null, req);

    res.status(201).json(tenant);
  } catch (error) {
    console.error('Error creating tenant:', error);
    if (error.code === 'ER_DUP_ENTRY') {
      res.status(409).json({
        error: {
          code: 'DOMAIN_EXISTS',
          message: 'Domain already exists'
        }
      });
    } else {
      res.status(500).json({
        error: {
          code: 'CREATE_ERROR',
          message: 'Failed to create tenant'
        }
      });
    }
  }
});

// Update tenant
app.put('/api/v1/admin/tenants/:tenantId', async (req, res) => {
  try {
    const { tenantId } = req.params;
    const updates = req.body;
    
    const tenant = await tenantDb.update(tenantId, updates, null, req);
    res.json(tenant);
  } catch (error) {
    console.error('Error updating tenant:', error);
    if (error.message === 'Tenant not found') {
      res.status(404).json({
        error: {
          code: 'TENANT_NOT_FOUND',
          message: 'Tenant not found'
        }
      });
    } else {
      res.status(500).json({
        error: {
          code: 'UPDATE_ERROR',
          message: 'Failed to update tenant'
        }
      });
    }
  }
});

// Delete tenant (soft delete)
app.delete('/api/v1/admin/tenants/:tenantId', async (req, res) => {
  try {
    const { tenantId } = req.params;
    await tenantDb.delete(tenantId, null, req);
    res.json({ success: true });
  } catch (error) {
    console.error('Error deleting tenant:', error);
    res.status(500).json({
      error: {
        code: 'DELETE_ERROR',
        message: 'Failed to delete tenant'
      }
    });
  }
});

// =============================================================================
// USER MANAGEMENT ENDPOINTS (Tenant-scoped)
// =============================================================================

// Get users for current tenant
app.get('/api/v1/users', authenticateToken, async (req, res) => {
  try {
    const { page = 1, limit = 10, search = '', role = '' } = req.query;
    const result = await userDb.getByTenant(
      req.user.tenantId, 
      parseInt(page), 
      parseInt(limit), 
      search, 
      role
    );
    
    res.json(result);
  } catch (error) {
    console.error('Error fetching users:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch users'
      }
    });
  }
});

// Get specific user
app.get('/api/v1/users/:userId', authenticateToken, async (req, res) => {
  try {
    const { userId } = req.params;
    const user = await userDb.getById(req.user.tenantId, userId);
    
    if (!user) {
      return res.status(404).json({
        error: {
          code: 'USER_NOT_FOUND',
          message: 'User not found'
        }
      });
    }
    
    res.json(user);
  } catch (error) {
    console.error('Error fetching user:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch user'
      }
    });
  }
});

// Create new user (admin only)
app.post('/api/v1/users', authenticateToken, requireRole(['admin']), checkUsageLimit('users'), async (req, res) => {
  try {
    const { email, name, role = 'user', metadata = {} } = req.body;
    
    if (!email || !name) {
      return res.status(400).json({
        error: {
          code: 'MISSING_FIELDS',
          message: 'Email and name are required'
        }
      });
    }

    const userId = `user_${uuidv4().replace(/-/g, '').substring(0, 12)}`;
    
    const user = await userDb.create({
      id: userId,
      tenant_id: req.user.tenantId,
      email,
      name,
      role,
      metadata
    }, req.user.userId, req);

    // Track user creation
    await subscriptionManager.trackUsage(req.user.tenantId, 'users', userId);

    res.status(201).json(user);
  } catch (error) {
    console.error('Error creating user:', error);
    if (error.code === 'ER_DUP_ENTRY') {
      res.status(409).json({
        error: {
          code: 'EMAIL_EXISTS',
          message: 'Email already exists in this tenant'
        }
      });
    } else {
      res.status(500).json({
        error: {
          code: 'CREATE_ERROR',
          message: 'Failed to create user'
        }
      });
    }
  }
});

// Update user
app.put('/api/v1/users/:userId', authenticateToken, async (req, res) => {
  try {
    const { userId } = req.params;
    const updates = req.body;
    
    // Users can only update themselves unless they're admin
    if (userId !== req.user.userId && !req.user.roles.includes('admin')) {
      return res.status(403).json({
        error: {
          code: 'INSUFFICIENT_PERMISSIONS',
          message: 'Can only update your own profile'
        }
      });
    }
    
    // Non-admins cannot change roles
    if (updates.role && !req.user.roles.includes('admin')) {
      delete updates.role;
    }
    
    const user = await userDb.update(req.user.tenantId, userId, updates, req.user.userId, req);
    res.json(user);
  } catch (error) {
    console.error('Error updating user:', error);
    if (error.message === 'User not found') {
      res.status(404).json({
        error: {
          code: 'USER_NOT_FOUND',
          message: 'User not found'
        }
      });
    } else {
      res.status(500).json({
        error: {
          code: 'UPDATE_ERROR',
          message: 'Failed to update user'
        }
      });
    }
  }
});

// Delete user (admin only)
app.delete('/api/v1/users/:userId', authenticateToken, requireRole(['admin']), async (req, res) => {
  try {
    const { userId } = req.params;
    
    // Prevent self-deletion
    if (userId === req.user.userId) {
      return res.status(400).json({
        error: {
          code: 'CANNOT_DELETE_SELF',
          message: 'Cannot delete your own account'
        }
      });
    }
    
    await userDb.delete(req.user.tenantId, userId, req.user.userId, req);
    res.json({ success: true });
  } catch (error) {
    console.error('Error deleting user:', error);
    if (error.message === 'User not found') {
      res.status(404).json({
        error: {
          code: 'USER_NOT_FOUND',
          message: 'User not found'
        }
      });
    } else {
      res.status(500).json({
        error: {
          code: 'DELETE_ERROR',
          message: 'Failed to delete user'
        }
      });
    }
  }
});

// Get user statistics for current tenant
app.get('/api/v1/users/stats', authenticateToken, requireRole(['admin']), async (req, res) => {
  try {
    const stats = await userDb.getStats(req.user.tenantId);
    res.json(stats);
  } catch (error) {
    console.error('Error fetching user stats:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch user statistics'
      }
    });
  }
});

// =============================================================================
// TENANT INFO ENDPOINTS
// =============================================================================

// Get current tenant info
app.get('/api/v1/tenant', authenticateToken, async (req, res) => {
  try {
    const tenant = await tenantDb.getById(req.user.tenantId);
    res.json(tenant);
  } catch (error) {
    console.error('Error fetching tenant info:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch tenant information'
      }
    });
  }
});

// Update current tenant (admin only)
app.put('/api/v1/tenant', authenticateToken, requireRole(['admin']), async (req, res) => {
  try {
    const updates = req.body;
    // Prevent changing critical fields
    delete updates.id;
    delete updates.domain;
    delete updates.status;
    
    const tenant = await tenantDb.update(req.user.tenantId, updates, req.user.userId, req);
    res.json(tenant);
  } catch (error) {
    console.error('Error updating tenant:', error);
    res.status(500).json({
      error: {
        code: 'UPDATE_ERROR',
        message: 'Failed to update tenant'
      }
    });
  }
});

// =============================================================================
// HEALTH AND UTILITY ENDPOINTS
// =============================================================================

// Health check with database connectivity
app.get('/health', async (req, res) => {
  try {
    const dbHealth = await healthCheck();
    res.json({
      status: 'ok',
      timestamp: new Date().toISOString(),
      database: dbHealth,
      version: '1.0.0'
    });
  } catch (error) {
    res.status(500).json({
      status: 'error',
      timestamp: new Date().toISOString(),
      error: error.message
    });
  }
});

// Get available tenants for login (public endpoint)
app.get('/api/v1/public/tenants', async (req, res) => {
  try {
    const result = await tenantDb.getAll(1, 100);
    const publicTenants = result.tenants
      .filter(t => t.status === 'active')
      .map(t => ({
        id: t.id,
        name: t.name,
        domain: t.domain,
        plan: t.plan
      }));
    
    res.json(publicTenants);
  } catch (error) {
    console.error('Error fetching public tenants:', error);
    res.status(500).json({
      error: {
        code: 'FETCH_ERROR',
        message: 'Failed to fetch tenants'
      }
    });
  }
});

// Error handling middleware
app.use((error, req, res, next) => {
  console.error('Unhandled error:', error);
  res.status(500).json({
    error: {
      code: 'INTERNAL_ERROR',
      message: 'Internal server error'
    }
  });
});

// 404 handler
app.use((req, res) => {
  res.status(404).json({
    error: {
      code: 'NOT_FOUND',
      message: 'Endpoint not found'
    }
  });
});

// Initialize database and start server
const startServer = async () => {
  try {
    await initDatabase();
    await initSubscriptionTables();
    await createDefaultSubscriptions();
    
    app.listen(PORT, () => {
      console.log(`🚀 SaaS Server running on http://localhost:${PORT}`);
      console.log(`📋 API Documentation:`);
      console.log(`   Health: GET /health`);
      console.log(`   Public Tenants: GET /api/v1/public/tenants`);
      console.log(`   Subscription Plans: GET /api/v1/plans`);
      console.log(`   Current Subscription: GET /api/v1/subscription`);
      console.log(`   Usage Stats: GET /api/v1/usage`);
      console.log(`   Change Plan: POST /api/v1/subscription/change`);
      console.log(`   Analytics: GET /api/v1/analytics (Premium+)`);
      console.log(`   Teams: GET /api/v1/teams (Basic+)`);
      console.log(`   File Upload: POST /api/v1/upload (with limits)`);
      console.log(`   Tenant Management: /api/v1/admin/tenants`);
      console.log(`   User Management: /api/v1/users`);
      console.log(`   Tenant Info: /api/v1/tenant`);
      console.log(`\n💡 Database: MySQL with full CRUD operations`);
      console.log(`🔐 Authentication: JWT with tenant isolation`);
      console.log(`💳 Subscriptions: Plan-based feature access & usage limits`);
      console.log(`📊 Features: Audit logging, pagination, search, role-based access`);
    });
  } catch (error) {
    console.error('❌ Failed to start server:', error.message);
    console.log('\n📋 Setup Instructions:');
    console.log('1. Install MySQL Server');
    console.log('2. Create database: CREATE DATABASE saas_app;');
    console.log('3. Update .env with your MySQL credentials');
    console.log('4. Run: npm run init-db');
    console.log('\nSee MYSQL_SETUP.md for detailed instructions');
    process.exit(1);
  }
};

// Graceful shutdown
process.on('SIGTERM', () => {
  console.log('SIGTERM received, shutting down gracefully');
  process.exit(0);
});

process.on('SIGINT', () => {
  console.log('SIGINT received, shutting down gracefully');
  process.exit(0);
});

startServer();